+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package grepros
 0.2.1 (2021-11-07)
 -------------------
 * add --write-format option, with HTML support
+* add --wrap-width option
 * add --order-bag-by option
 * handle topics with multiple message types
 * fix mixing up ROS2 bag start-end timestamps